+++ apps/gemebuild/src/i18n-pages/terms-and-conditions/fr.tsx
@@ -1,5 +1,6 @@
 import { TermsAndConditions, ITermsAndConditionsProps } from './components'
-import { ETabsId, SiteHelpTabs } from '../../components/tabs'
+import { SiteHelpTabs } from '../../components/tabs'
+import { ETabsId } from '../../components/tabs/type'
 import { getWarrantyTabsTextInfo } from '../../helpers/warranty-tabs-text-info'
 
 // 静态页面的navBar内容 配置文件 Fr版
