+++ apps/gemebuild/src/i18n-pages/warranty-extend/de.tsx
@@ -1,5 +1,6 @@
 import { IWarrantyExtendProps, WarrantyExtend } from './components'
-import { SiteHelpTabs, ETabsId } from '../../components/tabs'
+import { SiteHelpTabs } from '../../components/tabs'
+import { ETabsId } from '../../components/tabs/type'
 import { getWarrantyTabsTextInfo } from '../../helpers/warranty-tabs-text-info'
 
 // 静态页面的navBar内容 配置文件 De版
