+++ apps/gemebuild/src/i18n-pages/warranty/fr.tsx
@@ -1,9 +1,7 @@
 import { Warranty } from './components'
 import type { IWarrantyProps } from './components'
-import {
-  SiteHelpTabs,
-  ETabsId,
-} from '../../components/tabs'
+import { SiteHelpTabs } from '../../components/tabs'
+import { ETabsId } from '../../components/tabs/type'
 import { getWarrantyTabsTextInfo } from '../../helpers/warranty-tabs-text-info'
 
 // 静态页面的navbar内容 配置文件 Fr版
